Update photodir-import.py
In this modified version, the script checks if a command line argument is provided when running the script. The directory path is expected to be provided as the first argument after the script name.

If no argument is provided, it will display an error message with the correct usage format. If an argument is provided, it will call the import_photos() function with the provided directory path.